--- pint_output.log
@@ -1,5 +1,0 @@
-
-                         
-  No dirty files found.  
-                         
-
